requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
pytesseract
requests
python-dotenv
routers
supabase
pydantic
bcrypt
python-jose
pyjwt